split query hooks
