+++ LookdownComponent/Lookdown/components/LookdownControlClassic.tsx
@@ -15,11 +15,13 @@ import {
 } from "@fluentui/react";
 import { useQueryClient } from "@tanstack/react-query";
 import React, { useCallback, useEffect } from "react";
-import { useEntityOptions, useLanguagePack, useMetadata } from "../services/DataverseService";
 import { getClassicDropdownOptions } from "../services/DropdownHelper";
 import { getCustomFilterString, getHandlebarsVariables } from "../services/TemplateService";
 import { EntityOption, LookdownControlProps, OpenRecordMode } from "../types/typings";
 import { useAttributeOnChange } from "../hooks/useAttributeOnChange";
+import { useEntityOptions } from "../hooks/queries/useEntityOptions";
+import { useLanguagePack } from "../hooks/queries/useLanguagePack";
+import { useMetadata } from "../hooks/queries/useMetadata";
 
 const DEFAULT_BORDER_STYLES: IStyle = {
   borderColor: "#666",
