+++ LookdownComponent/Lookdown/services/DataverseService.ts
@@ -33,82 +33,6 @@ const DEFAULT_HEADERS = {
     'odata.include-annotations="OData.Community.Display.V1.FormattedValue,Microsoft.Dynamics.CRM.associatednavigationproperty,Microsoft.Dynamics.CRM.lookuplogicalname"',
 };
 
-export function useMetadata(lookupTable: string, lookupViewId: string) {
-  return useQuery({
-    queryKey: ["metadata", lookupTable, lookupViewId],
-    queryFn: () => getMetadata(lookupTable, lookupViewId),
-    enabled: !!lookupTable && !!lookupViewId,
-  });
-}
-
-export function useEntityOptions(
-  metadata: IMetadata | undefined,
-  customFilter?: string | null,
-  groupBy?: string | null,
-  optionTemplate?: string | null,
-  selectedItemTemplate?: string | null,
-  iconOptions?: ShowIconOptions,
-  iconSize?: IconSizes
-) {
-  const entitySetName = metadata?.lookupEntity.EntitySetName ?? "";
-  const lookupViewFetchXml = metadata?.lookupView.fetchxml ?? "";
-
-  const entityIcon =
-    metadata?.lookupEntity.IconVectorName ??
-    (iconSize === IconSizes.Large
-      ? metadata?.lookupEntity.IconMediumName ?? metadata?.lookupEntity.IconSmallName
-      : metadata?.lookupEntity.IconSmallName);
-
-  const recordImageUrlTemplate = metadata?.lookupEntity.RecordImageUrlTemplate;
-
-  let iconTemplate = "";
-  if (iconOptions === ShowIconOptions.RecordImage) {
-    iconTemplate = recordImageUrlTemplate ?? "";
-  } else if (iconOptions === ShowIconOptions.EntityIcon) {
-    iconTemplate = entityIcon ?? "";
-  }
-
-  return useQuery({
-    queryKey: [
-      "entityRecords",
-      entitySetName,
-      lookupViewFetchXml,
-      customFilter,
-      groupBy,
-      optionTemplate,
-      selectedItemTemplate,
-    ],
-    queryFn: () => {
-      const templateColumns: string[] = [];
-      if (optionTemplate || selectedItemTemplate) {
-        templateColumns.push(...getHandlebarsVariables(optionTemplate ?? "" + " " + selectedItemTemplate ?? ""));
-      }
-      const populatedFetchXml = getFetchTemplateString(lookupViewFetchXml ?? "", customFilter, templateColumns);
-      return getEntityRecords(
-        entitySetName,
-        metadata?.lookupEntity.PrimaryIdAttribute ?? "",
-        metadata?.lookupEntity.PrimaryNameAttribute ?? "",
-        populatedFetchXml,
-        groupBy,
-        optionTemplate,
-        selectedItemTemplate,
-        iconOptions,
-        iconTemplate,
-        iconSize
-      );
-    },
-    enabled: !!entitySetName && !!lookupViewFetchXml,
-  });
-}
-
-export function useLanguagePack(webResourcePath: string | undefined, defaultLanguagePack: LanguagePack) {
-  return useQuery({
-    queryKey: ["languagePack", webResourcePath],
-    queryFn: () => getLanguagePack(webResourcePath, defaultLanguagePack),
-    enabled: !!webResourcePath,
-  });
-}
-
 export function getLanguagePack(
   webResourceUrl: string | undefined,
   defaultLanguagePack: LanguagePack
@@ -140,7 +64,7 @@ export function getLanguagePack(
     .catch(() => languagePack);
 }
 
-async function getMetadata(lookupTable: string, lookupViewId: string): Promise<IMetadata> {
+export async function getMetadata(lookupTable: string, lookupViewId: string): Promise<IMetadata> {
   const [lookupEntity, lookupView] = await Promise.all([
     getEntityDefinition(lookupTable),
     getViewDefinition(lookupTable, lookupViewId),
@@ -235,7 +159,7 @@ export async function getLookupViewDefinition(entityName: string) {
   return view;
 }
 
-async function getEntityRecords(
+export async function getEntityRecords(
   entitySetName: string,
   primaryIdAttribute: string,
   primaryNameAttribute: string,
